Add additional information for branded foods
Fix a huge performance bug in the userFoodLoader. It was running the
entire list comprehension search for nutrients for every single food,
which in the case of branded foods meant traversing the entire 50+
million nutrient list 2.6 million times. Turned it into a dict and now
it is stupid fast (comparatively).