Add date into the food search UI
Very useful for branded foods where the name of the same food seems to
change every time they add it to the DB. so you get 6 almost identical
listings for a product, but since the names are somewhat different, we
can't just throw them out. The Date lets you figure out which one is the
most recent.

Also fix the issue with some foundation foods not having calorie values
because they populated a different nutrient id.

Update measures to things like "1 Bar" now become "Bar". Not messing
with trying to parse stuff like "2 Bars" because that potentially gets a
bit more messy